chore: standardise types to float/stdint, keep double where precision matters
- Default scaffold state var: double → float (0.0f)
- _config.py docstring examples: double/int → float/int32_t
- Examples: alpha, biquad coeffs/state, EMA prev: double → float
- array_processing: return_type int → int32_t; alpha double → float
- sliding_power: pos int → uint32_t; sum_sq stays double (precision demo)
- running_stats: mean/m2/min_val/max_val stay double (Welford precision)
- Docs: customization.md, template.md, examples/*.md updated to match
- test_new/test_cli: update default-gain assertion double → float

+++ src/just_makeit/examples/running_stats/.steps/05_demo.c
@@ -6,14 +6,14 @@
 int main(void) {
     running_stats_state_t *s = running_stats_create(0, 0.0, 0.0);
 
-    double        data[] = {2, 4, 4, 4, 5, 5, 7, 9};
+    float         data[] = {2, 4, 4, 4, 5, 5, 7, 9};
     float complex y;
     for (int i = 0; i < 8; i++)
-        y = running_stats_step(s, (float)data[i] + 0.0f * I);
+        y = running_stats_step(s, data[i] + 0.0f * I);
 
-    printf("n:        %d\n", running_stats_get_n(s));      /* 8     */
-    printf("mean:     %.4f\n", running_stats_get_mean(s)); /* 5.0000 */
-    printf("variance: %.4f\n", (double)cimagf(y));         /* 4.0000 */
+    printf("n:        %d\n", running_stats_get_n(s));        /* 8     */
+    printf("mean:     %.4f\n", running_stats_get_mean(s));   /* 5.0000 */
+    printf("variance: %.4f\n", (double)cimagf(y));           /* 4.0000 */
 
     running_stats_reset(s);
     printf("after reset: n=%d mean=%.1f\n", running_stats_get_n(s),
